feat: support interceptors arrays
This commit allows passing interceptors as single functions (current behaviour), or arrays of functions (new behaviour). This makes it much easier to execute multiple functions on the consumer side. The alternative would be to compose all functions into one, but that's more complicated in my opinion.

With this change, the following becomes possible.

```ts
$fetch('/api/endpoint', {
  onRequest: [
    () => { /* Do something */ },
    () => { /* Do something else */ },
  ],
  onResponse: [
    () => { /* Do something */ },
    () => { /* Do something else */ },
  ],
})
```

Note that calling `$fetch.create()` and then passing interceptors when calling `$fetch()` will keep overriding the interceptors as the current behaviour. I did not want to go too much into there as I'm not sure what would be the expected behaviour. Plus, continuing to override the interceptors means that this PR is not a breaking change.